Wealixir-client-portal/backend/requirements.txt
fastapi
uvicorn[standard]
starlette
python-multipart
python-dotenv
pydantic[email]
openpyxl
fpdf2
python-docx
requests
itsdangerous
